apps/hosted-agent/requirements.txt
azure-ai-agentserver-agentframework>=1.0.0b10
python-dotenv
azure-identity>=1.19.0
typing_extensions>=4.0.0
